First commit
The Monkey_single algorithm designed to simulate the Inifinite Monkey Theorem but if there were only one monkey.
The Monkey_threaded is designed to do the same thing but using threads to add more monkeys into the mix.
there will also be a distributed one for use with raspberry pis.